+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.10 - 2024-07-08
+* add plots to quality control for combo data
+
 ## gDRplots 0.0.9 - 2024-07-03
 * add plots to quality control of distribution of normalized data and metric stat